pyproject.toml
[tool.poetry]
name = "my_houdini_project"
version = "0.1.0"
description = "A Python project using Houdini's hython interpreter"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
python = "^3.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]
# Add any development dependencies you need here

[tool.poetry.scripts]
# Add your custom scripts here

[tool.poetry.dependencies.houdini]
path = "C:/Program Files/Side Effects Software/Houdini 19.5.682/bin/hython3.9.exe"
